Hat gerad mit einer accuracy von 78 Prozent abgeschlossen aber optimal ist das immer noch nicht :(

+++ src/networking/filter/Filter.java
@@ -10,7 +10,7 @@ public abstract class Filter {
 
     public abstract void backPropagate();
 
-    public abstract void applyChange(float learningRate);
+    public abstract void applyChange(double learningRate);
 
     public void zeroLoss() {}
 }
